requirements.txt
requests==2.21.0
SQLAlchemy==1.0.11
onetimepass==0.2.2
pyqrcode==1.2.1
redis==3.2.0
flask-bcrypt
flask-user
flask-sqlalchemy==2.1.0
pymysql
celery_once
hvac
pytest
pandas
pycryptodome